# add profiler support to mitigate performance investigation