Useful to get an idea of how long it takes from one point in the code to another point.
To use it, choose an id and call `logger.profile(your-id)` before and after the code you want to measure. e.g.
```
logger.profile('loadingComponent');
consumer.loadComponent(id);
logger.profile('loadingComponent');
```
Once done, the log writes the time it took to execute the code between the two calls.
If this is a repeated code it also shows how long this code was executed in total.
an example of the output:
```
[2020-12-04 16:24:46.100 -0500] INFO	 (31641): loadingComponent: 14ms. (total repeating 14ms)
[2020-12-04 16:24:46.110 -0500] INFO	 (31641): loadingComponent: 18ms. (total repeating 32ms)
  ```